requirements.txt
Pillow
PySimpleGUI